requirements.txt
# general libraries
requests
pillow
simpleaudio

# my libraries
nikki-utils

# google api + auth
google-auth-httplib2
google-auth-oauthlib
google-api-python-client

# aws
boto3

# testing
pytest-mock
pytest-cov